2-interface/checking type method moved to common class train; rubocop
