+++ .rubocop.yml
@@ -8,7 +8,7 @@ Style/FrozenStringLiteralComment:
   Enabled: false
 
 Layout/LineLength:
-  Max: 120
+  Max: 140
 
 Style/Documentation:
   Enabled: false
