+++ cargo_train.rb
@@ -1,0 +1,10 @@
+class CargoTrain < Train
+  def initialize(number)
+    super(number)
+    @type = :cargo
+  end
+
+  def add_wagon(wagon)
+    super(wagon) if suitable_wagon?(wagon)
+  end
+end
