+++ cargo_wagon.rb
@@ -1,6 +1,6 @@
 class CargoWagon
   attr_reader :type
-  
+
   def initialize
     @type = :cargo
   end
