--- passenger-wagon.rb
@@ -1,7 +1,0 @@
-class PassengerWagon
-    attr_reader :type
-
-    def initialize
-        @type = :passenger
-    end
-end
