+++ passenger_train.rb
@@ -1,0 +1,10 @@
+class PassengerTrain < Train
+  def initialize(number)
+    super(number)
+    @type = :passenger
+  end
+
+  def add_wagon(wagon)
+    super(wagon) if suitable_wagon?(wagon)
+  end
+end
