+++ passenger_wagon.rb
@@ -1,0 +1,7 @@
+class PassengerWagon
+  attr_reader :type
+
+  def initialize
+    @type = :passenger
+  end
+end
